refactor: extract function to achieve modularization
