+++ src/interface.ts
@@ -1,0 +1,6 @@
+export interface ElementDetail {
+    tagName: string | undefined;
+    id: string | undefined;
+    className: string | undefined;
+    textContent: string | undefined;
+}
